Created the application logic and unit tests (TDD)
Two ActionListeners implemented for the buttons. Each button is listened by a separate listener which handles the event to convert either Roman or integer. The listener calls the methods in the RomanNumeralParser and causes the result to be shown.

I'm not sure if it is good design to give the ActionListener access to the actual UI components but for this project it works well